requirements/prod.txt
# Django & DRF
Django>=4.0,<4.1
django-cors-headers>=3.11.0,<3.12
djangorestframework>=3.14.0,<3.15.0
djangorestframework-simplejwt>=5.2.2,<5.3.0
psycopg2-binary>=2.9.5,<3.0.0

# Deploying
gunicorn>=20.1.0,<21.0.0

# OpenAPI - Docs
drf-yasg>=1.21.4,<1.22

# Python
python-dotenv>=0.21.0,<0.22.0